+++ Sadie.Networking.Writers/Players/Other/PlayerDataWriter.cs
@@ -10,16 +10,16 @@ public class PlayerDataWriter : NetworkPacketWriter
     public PlayerDataWriter(IPlayerData playerData)
     {
         WriteShort(ServerPacketId.PlayerData);
-        WriteLong(playerData.Id);
+        WriteInt(playerData.Id);
         WriteString(playerData.Username);
         WriteString(playerData.FigureCode);
         WriteString(playerData.Gender == AvatarGender.Male ? "M" : "F");
         WriteString(playerData.Motto);
         WriteString(playerData.Username);
         WriteBoolean(false);
-        WriteLong(playerData.RespectsReceived);
-        WriteLong(playerData.RespectPoints);
-        WriteLong(playerData.RespectPointsPet);
+        WriteInt(playerData.RespectsReceived);
+        WriteInt(playerData.RespectPoints);
+        WriteInt(playerData.RespectPointsPet);
         WriteBoolean(false);
         WriteString(playerData.LastOnline.ToString());
         WriteBoolean(false);
